Move zk test to a different file to avoid merge conflicts
