+++ crates/forge/tests/cli/build.rs
@@ -1,8 +1,7 @@
 use crate::utils::generate_large_contract;
 use foundry_config::Config;
-use foundry_test_utils::{forgetest, snapbox::IntoData, str, util::OutputExt};
+use foundry_test_utils::{forgetest, snapbox::IntoData, str};
 use globset::Glob;
-use regex::Regex;
 
 forgetest_init!(can_parse_build_filters, |prj, cmd| {
     prj.clear();
@@ -164,17 +163,6 @@ forgetest_init!(build_sizes_no_forge_std, |prj, cmd| {
 "#]]
         .is_json(),
     );
-});
-
-// tests build output is as expected in zksync mode
-forgetest_init!(test_zk_build_sizes, |prj, cmd| {
-    cmd.args(["build", "--sizes", "--zksync", "--evm-version", "shanghai"]);
-    let stdout = cmd.assert_success().get_output().stdout_lossy();
-    let pattern =
-        Regex::new(r"\|\s*Counter\s*\|\s*800\s*\|\s*800\s*\|\s*450,199\s*\|\s*450,199\s*\|")
-            .unwrap();
-
-    assert!(pattern.is_match(&stdout), "Unexpected size output:\n{stdout}");
 });
 
 // tests that skip key in config can be used to skip non-compilable contract
