+++ crates/forge/tests/cli/main.rs
@@ -26,3 +26,6 @@ mod verify;
 mod verify_bytecode;
 
 mod ext_integration;
+mod zk_build;
+mod zk_cmd;
+mod zk_script;
