+++ crates/forge/tests/cli/test_cmd.rs
@@ -1339,40 +1339,6 @@ Traces:
     ]]);
 });
 
-// Related to: https://github.com/matter-labs/foundry-zksync/issues/478
-forgetest_async!(test_zk_can_detect_call_to_empty_contract, |prj, cmd| {
-    foundry_test_utils::util::initialize(prj.root());
-
-    prj.add_test(
-        "CallEmptyCode.t.sol",
-        r#"
-import "forge-std/Test.sol";
-
-// https://github.com/matter-labs/foundry-zksync/issues/478
-contract CallEmptyCode is Test {
-    // This test should make our EraVM tracer print out an ERROR trace
-    function testFailDetectEmptyCodeContracts() external {
-        address mockMe = address(123456789);
-
-        vm.mockCall(mockMe, abi.encodeWithSignature("foo()"), abi.encode(42));
-
-        (bool success, bytes memory ret) = mockMe.call(abi.encodeWithSignature("bar()"));
-
-        require(success, "callMethod failed");
-        require(keccak256(ret) == keccak256(abi.encode(42)), "return not as expected");
-    }
-}
-"#,
-    )
-    .unwrap();
-    cmd.args(["test", "--zksync", "--evm-version", "shanghai", "--mc", "CallEmptyCode"]);
-
-    cmd.assert_success()
-        .get_output()
-        .stdout_lossy()
-        .contains("call may fail or behave unexpectedly due to empty code");
-});
-
 // tests that `forge test` with a seed produces deterministic random values for uint and addresses.
 forgetest_init!(deterministic_randomness_with_seed, |prj, cmd| {
     prj.wipe_contracts();
